refactor: extract useProviderNodeActions (#3501 Phase 1t.4)
